implementing TimeSlot
